seo-audit-tool/requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
tldextract>=5.0.0
pyyaml>=6.0.0
